+++ src/query/sql/src/planner/optimizer/heuristic/heuristic.rs
@@ -100,6 +100,8 @@ impl HeuristicOptimizer {
         let pre_optimized = self.pre_optimize(s_expr)?;
         let optimized = self.optimize_expression(&pre_optimized)?;
         let post_optimized = self.post_optimize(optimized)?;
+        // Post optimization may bring potential optimization opportunities
+        // So we can call `optimize_expression` again to optimize completely
         let final_s_expr = self.optimize_expression(&post_optimized)?;
 
         Ok(final_s_expr)
