+++ tests/logictest/suites/mode/standalone/explain/explain.test
@@ -532,6 +532,163 @@ UnionAll
     ├── push downs: [filters: [], limit: NONE]
     └── output columns: [0]
 
+statement query T
+explain select * from t1,t2 where (t1.a > 1 and t2.a > 2) or (t1.b < 3 and t2.b < 4);
+
+----
+Filter
+├── filters: [or(and(>(t1.a (#0), 1), >(t2.a (#2), 2)), and(<(t1.b (#1), 3), <(t2.b (#3), 4)))]
+└── HashJoin
+    ├── join type: CROSS
+    ├── build keys: []
+    ├── probe keys: []
+    ├── filters: []
+    ├── TableScan(Build)
+    │   ├── table: default.default.t2
+    │   ├── read rows: 0
+    │   ├── read bytes: 0
+    │   ├── partitions total: 0
+    │   ├── partitions scanned: 0
+    │   └── push downs: [filters: [((a > 2) or (b < 4))], limit: NONE]
+    └── TableScan(Probe)
+        ├── table: default.default.t1
+        ├── read rows: 0
+        ├── read bytes: 0
+        ├── partitions total: 0
+        ├── partitions scanned: 0
+        └── push downs: [filters: [((a > 1) or (b < 3))], limit: NONE]
+
+statement query T
+explain select * from t1,t2 where (t1.a > 1 and t2.a > 2) or (t1.b < 3 and t2.b < 4) or t1.a = 2;
+
+----
+Filter
+├── filters: [or(or(and(>(t1.a (#0), 1), >(t2.a (#2), 2)), and(<(t1.b (#1), 3), <(t2.b (#3), 4))), =(t1.a (#0), 2))]
+└── HashJoin
+    ├── join type: CROSS
+    ├── build keys: []
+    ├── probe keys: []
+    ├── filters: []
+    ├── TableScan(Build)
+    │   ├── table: default.default.t2
+    │   ├── read rows: 0
+    │   ├── read bytes: 0
+    │   ├── partitions total: 0
+    │   ├── partitions scanned: 0
+    │   └── push downs: [filters: [], limit: NONE]
+    └── TableScan(Probe)
+        ├── table: default.default.t1
+        ├── read rows: 0
+        ├── read bytes: 0
+        ├── partitions total: 0
+        ├── partitions scanned: 0
+        └── push downs: [filters: [((a > 1) or ((b < 3) or (a = 2)))], limit: NONE]
+
+statement ok
+create table t3(a int, b int);
+
+statement query T
+explain select * from t1,t2, t3 where (t1.a > 1 and t2.a > 2) or (t1.b < 3 and t2.b < 4) or t3.a = 2;
+
+----
+Filter
+├── filters: [or(or(and(>(t1.a (#0), 1), >(t2.a (#2), 2)), and(<(t1.b (#1), 3), <(t2.b (#3), 4))), =(t3.a (#4), 2))]
+└── HashJoin
+    ├── join type: CROSS
+    ├── build keys: []
+    ├── probe keys: []
+    ├── filters: []
+    ├── TableScan(Build)
+    │   ├── table: default.default.t3
+    │   ├── read rows: 0
+    │   ├── read bytes: 0
+    │   ├── partitions total: 0
+    │   ├── partitions scanned: 0
+    │   └── push downs: [filters: [], limit: NONE]
+    └── HashJoin(Probe)
+        ├── join type: CROSS
+        ├── build keys: []
+        ├── probe keys: []
+        ├── filters: []
+        ├── TableScan(Build)
+        │   ├── table: default.default.t2
+        │   ├── read rows: 0
+        │   ├── read bytes: 0
+        │   ├── partitions total: 0
+        │   ├── partitions scanned: 0
+        │   └── push downs: [filters: [], limit: NONE]
+        └── TableScan(Probe)
+            ├── table: default.default.t1
+            ├── read rows: 0
+            ├── read bytes: 0
+            ├── partitions total: 0
+            ├── partitions scanned: 0
+            └── push downs: [filters: [], limit: NONE]
+
+statement query T
+explain select * from t1,t2, t3 where ((t1.a > 1 and t2.a > 2) or (t1.b < 3 and t2.b < 4)) and t3.a > 1;
+
+----
+HashJoin
+├── join type: CROSS
+├── build keys: []
+├── probe keys: []
+├── filters: []
+├── TableScan(Build)
+│   ├── table: default.default.t3
+│   ├── read rows: 0
+│   ├── read bytes: 0
+│   ├── partitions total: 0
+│   ├── partitions scanned: 0
+│   └── push downs: [filters: [(a > 1)], limit: NONE]
+└── Filter(Probe)
+    ├── filters: [or(and(>(t1.a (#0), 1), >(t2.a (#2), 2)), and(<(t1.b (#1), 3), <(t2.b (#3), 4)))]
+    └── HashJoin
+        ├── join type: CROSS
+        ├── build keys: []
+        ├── probe keys: []
+        ├── filters: []
+        ├── TableScan(Build)
+        │   ├── table: default.default.t2
+        │   ├── read rows: 0
+        │   ├── read bytes: 0
+        │   ├── partitions total: 0
+        │   ├── partitions scanned: 0
+        │   └── push downs: [filters: [((a > 2) or (b < 4))], limit: NONE]
+        └── TableScan(Probe)
+            ├── table: default.default.t1
+            ├── read rows: 0
+            ├── read bytes: 0
+            ├── partitions total: 0
+            ├── partitions scanned: 0
+            └── push downs: [filters: [((a > 1) or (b < 3))], limit: NONE]
+
+statement query T
+explain select * from t1,t2 where ((t1.a > 1 or t1.b < 2) and t2.a > 2) or (t1.b < 3 and t2.b < 4);
+
+----
+Filter
+├── filters: [or(and(or(>(t1.a (#0), 1), <(t1.b (#1), 2)), >(t2.a (#2), 2)), and(<(t1.b (#1), 3), <(t2.b (#3), 4)))]
+└── HashJoin
+    ├── join type: CROSS
+    ├── build keys: []
+    ├── probe keys: []
+    ├── filters: []
+    ├── TableScan(Build)
+    │   ├── table: default.default.t2
+    │   ├── read rows: 0
+    │   ├── read bytes: 0
+    │   ├── partitions total: 0
+    │   ├── partitions scanned: 0
+    │   └── push downs: [filters: [((a > 2) or (b < 4))], limit: NONE]
+    └── TableScan(Probe)
+        ├── table: default.default.t1
+        ├── read rows: 0
+        ├── read bytes: 0
+        ├── partitions total: 0
+        ├── partitions scanned: 0
+        └── push downs: [filters: [(((a > 1) or (b < 2)) or (b < 3))], limit: NONE]
+
 statement ok
 drop table t1;
